Integrate formatting with Nova's onWillSave
Before this change, it was easy to get stuck in a loop where a document
was saved, _then_ formatted and overwritten, which then marked the
document as unsaved.

Nova's onWillSave function when used with javascript Promises is
perfect for the common “format on save” functionality. The returned
Promise ensures that Nova waits at least 5 seconds while the formatter
runs. Once the Promise is resolved, the document saves with the newly
formatted text. If it rejects instead, the user is informed that
something went wrong.

https://docs.nova.app/api-reference/text-editor/#onwillsavecallback-this
value